+++ lib/src/views/image_fall_view.dart
@@ -1,4 +1,5 @@
 import 'dart:math';
+import 'dart:typed_data';
 
 import 'package:flutter/material.dart';
 import 'package:simple_3d/simple_3d.dart';
@@ -6,15 +7,17 @@ import 'package:util_simple_3d/util_simple_3d.dart';
 import '../../sakura_blizzard.dart';
 
 ///
-/// (en) A view widget of colorful cubes rotating and falling.
+/// (en) This is a view that gives the effect of the specified image falling.
 ///
-/// (ja) カラフルなキューブが回転しながら落下するビューです。
+/// (ja) 指定した画像が落下するビューです。
 ///
 /// Author Masahide Mori
 ///
-class ColorfulCubeView extends StatefulWidget {
+class ImageFallView extends StatefulWidget {
   final Widget child;
   final Size viewSize;
+  final List<Uint8List> images;
+  final List<Uint8List> backImages;
   final bool isRandomPositionY;
   final int frontObjNum;
   final int backObjNum;
@@ -27,6 +30,11 @@ class ColorfulCubeView extends StatefulWidget {
 
   /// * [child] : A child view will placed between the front and back layers.
   /// * [viewSize] : This view size.
+  /// * [images] : A list of images you want to drop.
+  /// If multiple images are specified,
+  /// the image to be used will be randomly selected during initialization.
+  /// * [backImages] : List of images on the back. Must be the same number and
+  /// in the same order as the surface.
   /// * [isRandomPositionY] : Specifies whether to randomize the y-coordinate
   /// at which the object starts falling.
   /// * [frontObjNum] : Number of front layer objects you want to generate.
@@ -44,22 +52,24 @@ class ColorfulCubeView extends StatefulWidget {
   /// and 1 means the brightness will not change.
   /// * [customPhysicsCreation] : You can create your own behavior and use it.
   /// If this is not null, the dropType parameter is ignored.
-  const ColorfulCubeView(
+  const ImageFallView(
       {required this.child,
       required this.viewSize,
+      required this.images,
+      required this.backImages,
       this.isRandomPositionY = true,
       this.frontObjNum = 20,
       this.backObjNum = 20,
       this.frontObjSize = const VRange(min: 8, max: 32),
       this.backObjSize = const VRange(min: 8, max: 32),
-      this.dropType = EnumDropType.spinDrop3D,
+      this.dropType = EnumDropType.hirahiraDrop,
       this.fps = 60,
       this.minBrightness = 0.0,
       this.customPhysicsCreation,
       super.key});
 
   @override
-  State<ColorfulCubeView> createState() => _ColorfulCubeViewState();
+  State<ImageFallView> createState() => _ImageFallViewState();
 
   /// Convert enum to physics object.
   Sp3dPhysics _getDropPhysics(int fps) {
@@ -78,26 +88,14 @@ class ColorfulCubeView extends StatefulWidget {
     }
   }
 
-  /// Return random color material.
-  Sp3dMaterial _getRandomColorMaterial() {
-    double src = Random().nextDouble();
-    if (src < 0.25) {
-      return FSp3dMaterial.red;
-    } else if (src >= 0.25 && src < 0.5) {
-      return FSp3dMaterial.green;
-    } else if (src >= 0.5 && src < 0.75) {
-      return FSp3dMaterial.blue;
-    } else {
-      return SakuraBlizzardMaterials.yellow;
-    }
-  }
-
   /// Creates an object with the specified size range.
   /// * [targetSize] : The size range of generation object.
   Sp3dObj createObj(VRange targetSize) {
     final double objSize = targetSize.getRandomInRange();
-    final Sp3dObj r = UtilSp3dGeometry.cube(objSize, objSize, objSize, 1, 1, 1,
-        material: _getRandomColorMaterial());
+    final Sp3dObj r = UtilImageTileCreator.imageTile(objSize);
+    final int targetImageIndex = Random().nextInt(images.length);
+    r.images.add(images[targetImageIndex]);
+    r.images.add(backImages[targetImageIndex]);
     r.physics = _getDropPhysics(fps);
     if (r.physics!.rotateAxis != null) {
       r.rotate(r.physics!.rotateAxis!, Random().nextDouble() * 360 * pi / 180);
@@ -112,7 +110,7 @@ class ColorfulCubeView extends StatefulWidget {
   }
 }
 
-class _ColorfulCubeViewState extends State<ColorfulCubeView> {
+class _ImageFallViewState extends State<ImageFallView> {
   final List<Sp3dObj> _frontObjs = [];
   final List<Sp3dObj> _backObjs = [];