+++ lib/src/views/sakura_blizzard_view.dart
@@ -24,6 +24,7 @@ class SakuraBlizzardView extends StatefulWidget {
   final EnumDropType dropType;
   final double minBrightness;
   final int fps;
+  final Sp3dPhysics Function()? customPhysicsCreation;
 
   /// * [child] : A child view will placed between the front and back layers.
   /// * [viewSize] : This view size.
@@ -42,6 +43,8 @@ class SakuraBlizzardView extends StatefulWidget {
   /// * [minBrightness] : The maximum magnification of how dark the object
   /// will be when rotated. 0 is the darkest,
   /// and 1 means the brightness will not change.
+  /// * [customPhysicsCreation] : You can create your own behavior and use it.
+  /// If this is not null, the dropType parameter is ignored.
   const SakuraBlizzardView(
       {required this.child,
       required this.viewSize,
@@ -53,6 +56,7 @@ class SakuraBlizzardView extends StatefulWidget {
       this.dropType = EnumDropType.hirahiraDrop,
       this.fps = 60,
       this.minBrightness = 0.93,
+      this.customPhysicsCreation,
       super.key});
 
   @override
@@ -60,6 +64,9 @@ class SakuraBlizzardView extends StatefulWidget {
 
   /// Convert enum to physics object.
   Sp3dPhysics _getDropPhysics(int fps) {
+    if (customPhysicsCreation != null) {
+      return customPhysicsCreation!();
+    }
     switch (dropType) {
       case EnumDropType.rotatingDrop:
         return RotatingDropPhysics(fps: fps);
@@ -67,6 +74,8 @@ class SakuraBlizzardView extends StatefulWidget {
         return SpinDrop3DPhysics(fps: fps);
       case EnumDropType.hirahiraDrop:
         return HirahiraDropPhysics(viewSize, fps: fps);
+      case EnumDropType.basicDrop:
+        return BasicDropPhysics();
     }
   }
 
@@ -77,7 +86,9 @@ class SakuraBlizzardView extends StatefulWidget {
     final Sp3dObj r =
         UtilSakuraCreator.sakuraPetal(objSize / 1.5, objSize, objSize / 6);
     r.physics = _getDropPhysics(fps);
-    r.rotate(r.physics!.rotateAxis!, Random().nextDouble() * 360 * pi / 180);
+    if (r.physics!.rotateAxis != null) {
+      r.rotate(r.physics!.rotateAxis!, Random().nextDouble() * 360 * pi / 180);
+    }
     if (isRandomPositionY) {
       r.move(Sp3dV3D(viewSize.width * Random().nextDouble(),
           (viewSize.height * 1.125) * Random().nextDouble(), 0));